added StringWriter interface in decoration package
