+++ internal/decoration/aurora.go
@@ -2,6 +2,7 @@ package decoration
 
 import (
 	"fmt"
+	"io"
 	"math"
 	"unicode"
 	"unicode/utf8"
@@ -30,6 +31,31 @@ func (w *Writer) writeAsAurora(b []byte) (nn int, err error) {
 		b = b[size:]
 	}
 
+	return w.writer.Write(w.buf.Bytes())
+}
+
+func (w *Writer) writeStringAsAurora(s string) (nn int, err error) {
+	defer w.buf.Reset()
+
+	for _, char := range s {
+		if char == '\n' {
+			w.buf.WriteRune(char)
+			continue
+		}
+		if unicode.IsSpace(char) {
+			w.buf.WriteRune(char)
+		} else {
+			fmt.Fprintf(&w.buf, "\033[38;5;%d%sm%c\033[0m",
+				rgb(float64(w.options.colorSeq)),
+				w.options.maybeBold(),
+				char,
+			)
+		}
+		w.options.colorSeq++
+	}
+	if sw, ok := w.writer.(io.StringWriter); ok {
+		return sw.WriteString(w.buf.String())
+	}
 	return w.writer.Write(w.buf.Bytes())
 }
