+++ internal/decoration/decoration.go
@@ -2,10 +2,7 @@ package decoration
 
 import (
 	"bytes"
-	"fmt"
 	"io"
-	"unicode"
-	"unicode/utf8"
 )
 
 type options struct {
@@ -54,6 +51,11 @@ type Writer struct {
 	options *options
 }
 
+var _ interface {
+	io.Writer
+	io.StringWriter
+} = (*Writer)(nil)
+
 // NewWriter creates a new writer.
 func NewWriter(w io.Writer, opts ...Option) *Writer {
 	options := new(options)
@@ -84,22 +86,18 @@ func (w *Writer) Write(b []byte) (nn int, err error) {
 	}
 }
 
-func (w *Writer) writeAsDefaultBold(b []byte) (nn int, err error) {
-	defer w.buf.Reset()
-
-	for len(b) > 0 {
-		char, size := utf8.DecodeRune(b)
-		if char == '\n' {
-			w.buf.WriteRune(char)
-			b = b[size:]
-			continue
+func (w *Writer) WriteString(s string) (n int, err error) {
+	switch {
+	case w.options.withAurora:
+		return w.writeStringAsAurora(s)
+	case w.options.withRainbow:
+		return w.writeStringAsRainbow(s)
+	case w.options.withBold:
+		return w.writeStringAsDefaultBold(s)
+	default:
+		if sw, ok := w.writer.(io.StringWriter); ok {
+			return sw.WriteString(w.buf.String())
 		}
-		if unicode.IsSpace(char) {
-			w.buf.WriteRune(char)
-		} else {
-			fmt.Fprintf(&w.buf, "\x1b[1m%c\x1b[0m", char)
-		}
-		b = b[size:]
+		return w.writer.Write([]byte(s))
 	}
-	return w.writer.Write(w.buf.Bytes())
 }
